refactor: 클라이언트 컴포넌트로 분리
