+++ src/app/(tabs)/home/_hooks/use-side-effects.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useRef } from "react";
 
 import { ev } from "@/lib/gtm";
